requirements.txt

streamlit>=1.36.0
pandas>=2.2.2
jinja2>=3.1.4
fpdf2>=2.7.9
